ch(project folders and files): Rearranges folders and files
This rearranges folders and files as instructed

[Deliveries #161463017]
