+++ APIs/__init__.py
@@ -3,4 +3,4 @@ from flask import Flask
 
 app = Flask(__name__)
 
-from APIs import routes
+import APIs.views
