+++ tests/test_app.py
@@ -2,7 +2,7 @@
 import unittest
 import json
 from APIs import app
- 
+
 class ManagerTestCase(unittest.TestCase):
 
     def setUp(self):
@@ -26,11 +26,12 @@ class ManagerTestCase(unittest.TestCase):
             "price": 200
         }
 
-        self.client = app.test_client(self)
+        self.client = app.test_client()
 
     def tearDown(self):
         self.app_user.clear()
         self.product.clear()
+        self.sale.clear()
 
     def test_create_user(self):
         resp = self.client.post(
@@ -85,6 +86,34 @@ class ManagerTestCase(unittest.TestCase):
         self.assertEqual(resp.status_code, 201)
         self.assertIn('pants', str(resp.data))
 
+    def test_create_product_empty_fields(self):
+        rp = self.client.post(
+            '/storemanager/api/v1/products',
+            data=json.dumps({
+            "prod_name": '',
+            "category": '',
+            "stock": '',
+            "min_stock": '',    
+            "price": ''}),
+            content_type='application/json'
+            )
+        self.assertEqual(rp.status_code, 400)
+
+    def test_create_product_number_input(self):
+        rp = self.client.post(
+            '/storemanager/api/v1/products',
+            data=json.dumps({
+            "prod_name": "bell_bottoms",
+            "category": "pants",
+            "stock": '20A',
+            "min_stock": '10P',    
+            "price": '200k'}),
+            content_type='application/json'
+            )
+        rp_decode = json.loads(rp.data.decode())
+        self.assertEqual(rp.status_code, 400)
+        self.assertIn('Numbers expected for units', rp_decode['message'])
+
     def test_update_product(self):
         rp = self.client.post(
             '/storemanager/api/v1/products',
@@ -105,12 +134,37 @@ class ManagerTestCase(unittest.TestCase):
         self.assertEqual(resp.status_code, 200)
         self.assertIn('300', str(resp.data))
 
+    def test_update_product_fields_input(self):
+        rp = self.client.post(
+            '/storemanager/api/v1/products',
+            data=json.dumps(self.product),
+            content_type='application/json'
+            )
+        self.assertEqual(rp.status_code, 201)
+        resp = self.client.put(
+            '/storemanager/api/v1/products/1',
+            data=json.dumps({
+                "prod_name": 20,
+                "category": 50,
+                "stock": '20',
+                "min_stock": '10',
+                "price": '200'
+                }),
+            content_type='application/json'
+            )
+        self.assertEqual(resp.status_code, 400)
+
+
     def test_view_a_product(self):
         resp = self.client.get('/storemanager/api/v1/products/1')
         self.assertEqual(resp.status_code, 200)
         self.assertIn('pants', str(resp.data))
         self.assertIn('id', str(resp.data))
 
+    def test_delete_product(self):
+        resp = self.client.delete('/storemanager/api/v1/products/1')
+        self.assertEqual(resp.status_code, 200)
+        self.assertIn('true', str(resp.data))
 
     def test_view_all_products(self):
         resp = self.client.get('/storemanager/api/v1/products')
@@ -118,10 +172,15 @@ class ManagerTestCase(unittest.TestCase):
         self.assertIn('pants', str(resp.data))
         self.assertIn('id', str(resp.data))      
 
-    def test_delete_product(self):
-        resp = self.client.delete('/storemanager/api/v1/products/1')
-        self.assertEqual(resp.status_code, 200)
-        self.assertIn('true', str(resp.data))
+    def test_view_all_products_empty_list(self):
+        self.client.delete(
+            '/storemanager/api/v1/products/1')
+        self.client.delete(
+            '/storemanager/api/v1/products/2')
+        rp = self.client.get('/storemanager/api/v1/products')
+        rp_decode = json.loads(rp.data.decode())
+        self.assertEqual(rp.status_code, 400)
+        self.assertIn('Inventory is empty!', rp_decode['message'])
 
 
     """
